fix: same SVG filenames in different directories (#69)
* fix: same SVG filenames in different directories

The `svgs` map in `SVGManager` now uses the `filePath` as the key instead of the `name`.
This ensures that SVGs with the same filename but in different directories are uniquely identified.

The `idify` function now returns only the name. The prefix is added separately.
The returned `id` is added to the `SvgMapObject` and used in place of the initial `name`.
It makes the use of the `prefix` more consistent everywhere.

This could be improved as, for now, without custom `idify` function,
having same SVG filenames in different directories will result in the same `id` in the spritemap.

This could also potentially break existing configurations that rely on the prefix being added in the `idify` function.

* Add documentation (#65)

---------

Co-authored-by: Thierry Michel <thmichel@gmail.com>